fix: :bug: fix due to merge and move type to model
